# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "rango服务端接口文档，注意模块名称",
        "title": "rango 服务接口 Swagger 文档",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:2020",
    "paths": {
...
        "/api/params/:id": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Api模块"
                ],
                "summary": "测试请求参数",
                "parameters": [
                    {
                        "description": "name入参，测试入参",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.Params"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "响应的body体",
                        "schema": {
                            "type": "map"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.Params": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}